resolve suggestions

+++ components/org.wso2.carbon.identity.oauth.par/src/main/java/org/wso2/carbon/identity/oauth/par/common/ParConstants.java
@@ -23,7 +23,7 @@ package org.wso2.carbon.identity.oauth.par.common;
  */
 public class ParConstants {
 
-    public static final long EXPIRES_IN_DEFAULT_VALUE = 60;
+    public static final int EXPIRES_IN_DEFAULT_VALUE = 60;
     public static final long SEC_TO_MILLISEC_FACTOR = 1000;
     public static final String UTC = "UTC";
     public static final String EXPIRES_IN = "expires_in";
